Move various util modules to util/
