+++ lib/identity.js
@@ -9,10 +9,10 @@ module.exports.loadTypeRegistry = loadTypeRegistry
 module.exports.isPrivateTypeIndex = isPrivateTypeIndex
 module.exports.isPublicTypeIndex = isPublicTypeIndex
 
-var graphUtil = require('./graph-util')
+var graphUtil = require('./util/graph-util')
 var webClient = require('./web')
-var SolidProfile = require('./solid-profile')
-var rdf = require('./rdf-parser').rdflib
+var SolidProfile = require('./solid/profile')
+var rdf = require('./util/rdf-parser').rdflib
 var Vocab = require('./vocab')
 
 /**
