+++ lib/solid/profile.js
@@ -1,12 +1,12 @@
 'use strict'
 /**
- * @module solid-profile
+ * @module profile
  */
 module.exports = SolidProfile
-var rdf = require('./rdf-parser').rdflib
-var Vocab = require('./vocab')
-var identity = require('./identity')
-var graphUtil = require('./graph-util')
+var rdf = require('../util/rdf-parser').rdflib
+var Vocab = require('../vocab')
+var identity = require('../identity')
+var graphUtil = require('../util/graph-util')
 
 /**
  * Provides convenience methods for a WebID Profile.
