+++ lib/solid/response.js
@@ -1,10 +1,10 @@
 'use strict'
 /**
-* @module solid-response
+* @module response
 */
 module.exports = SolidResponse
 
-var webUtil = require('./web-util')
+var webUtil = require('../util/web-util')
 
 /**
 * Provides a wrapper around an XHR response object, and adds several
