+++ lib/util/graph-util.js
