+++ lib/web.js
@@ -4,8 +4,8 @@
  * @module web
  */
 var config = require('../config')
-var graphUtil = require('./graph-util')
-var SolidResponse = require('./solid-response')
+var graphUtil = require('./util/graph-util')
+var SolidResponse = require('./solid/response')
 var Vocab = require('./vocab')
 var XMLHttpRequest = require('./xhr')
 
@@ -213,7 +213,7 @@ var SolidWebClient = {
     proxyUrl = proxyUrl || config.proxyUrl
     timeout = timeout || config.timeout
     if (config.parser === 'rdflib') {
-      var getParsedGraph = require('./web-rdflib').getParsedGraph
+      var getParsedGraph = require('./util/web-rdflib').getParsedGraph
     } else {
       throw Error('Parser library not supported: ' + config.parser)
     }
@@ -285,7 +285,7 @@ var SolidWebClient = {
    *     object if not successful)
    */
   patch: function patch (url, toDel, toIns) {
-    var composePatchQuery = require('./web-util').composePatchQuery
+    var composePatchQuery = require('./util/web-util').composePatchQuery
     var data = composePatchQuery(toDel, toIns)
     var mimeType = 'application/sparql-update'
     var options = {}
